+++ KVASS/settings.cs
@@ -12,7 +12,7 @@ namespace KVASSNS
     public class KVASSSimSettings : GameParameters.CustomParameterNode
     {
         public override string Title { get { return Localizer.Format("#KVASS_sim_title"); } }
-        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.CAREER | GameParameters.GameMode.SCIENCE; } }
+        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.NOTMISSION; } }
         public override string Section { get { return "KVASS"; } }
         public override string DisplaySection { get { return "KVASS"; } }
         public override int SectionOrder { get { return 1; } }
@@ -24,66 +24,80 @@ namespace KVASSNS
         [GameParameters.CustomIntParameterUI("#KVASS_sim_ignore_SPH")]
         public bool IgnoreSPH { get; private set; } = false;
 
+        [GameParameters.CustomStringParameterUI("SSDM", gameMode = GameParameters.GameMode.SANDBOX,
+            autoPersistance = true, lines = 4, title = "#KVASS_sim_sand_disabled")]
+        public string StringSSDM = "";
 
-        [GameParameters.CustomFloatParameterUI("#KVASS_sim_career_vessel", gameMode = GameParameters.GameMode.CAREER,
+        [GameParameters.CustomIntParameterUI("#KVASS_sim_rel_cost_enabled", gameMode = GameParameters.GameMode.CAREER)]
+        public bool RelCostEnabled { get; private set; } = true;
+
+        [GameParameters.CustomIntParameterUI("#KVASS_sim_rel_science_enabled", gameMode = GameParameters.GameMode.SCIENCE)]
+        public bool RelScienceEnabled { get; private set; } = true;
+
+        [GameParameters.CustomFloatParameterUI("#KVASS_sim_rel_cost", gameMode = GameParameters.GameMode.CAREER,
             minValue = 0.0f, maxValue = 20.0f, displayFormat = "N1")]
-        public double CareerVessel { get; private set; } = 2.0f;
+        public double RelCost { get; private set; } = 2.0f;
 
-        [GameParameters.CustomFloatParameterUI("#KVASS_sim_science_vessel", gameMode = GameParameters.GameMode.SCIENCE,
+        [GameParameters.CustomFloatParameterUI("#KVASS_sim_rel_science", gameMode = GameParameters.GameMode.SCIENCE,
             minValue = 0.0f, maxValue = 20.0f, displayFormat = "N1")]
-        public float ScienceVessel { get; private set; } = 1.0f;
+        public float RelScience { get; private set; } = 1.0f;
 
 
-        [GameParameters.CustomParameterUI("#KVASS_sim_career_addFee", toolTip = "#KVASS_sim_career_addFee_tooltip",
+        [GameParameters.CustomParameterUI("#KVASS_sim_const_cost_enabled",
             gameMode = GameParameters.GameMode.CAREER)]
-        public bool CareerIsConst { get; private set; } = false;
+        public bool ConstCostEnabled { get; private set; } = false;
 
-        [GameParameters.CustomParameterUI("#KVASS_sim_science_addFee", toolTip = "#KVASS_sim_science_addFee_tooltip",
+        [GameParameters.CustomParameterUI("#KVASS_sim_const_science_enabled",
             gameMode = GameParameters.GameMode.SCIENCE)]
-        public bool ScienceIsConst { get; private set; } = false;
+        public bool ConstScienceEnabled { get; private set; } = false;
 
-
-        [GameParameters.CustomIntParameterUI("#KVASS_sim_career_const", gameMode = GameParameters.GameMode.CAREER,
+        [GameParameters.CustomIntParameterUI("#KVASS_sim_const_cost", gameMode = GameParameters.GameMode.CAREER,
             minValue = 500, maxValue = 100000, stepSize = 500)]
-        public int CareerConst { get; private set; } = 1000;
+        public int ConstCost { get; private set; } = 1000;
 
-        [GameParameters.CustomFloatParameterUI("#KVASS_sim_science_const", gameMode = GameParameters.GameMode.SCIENCE,
+        [GameParameters.CustomFloatParameterUI("#KVASS_sim_const_science", gameMode = GameParameters.GameMode.SCIENCE,
             minValue = 0.1f, maxValue = 20.0f, displayFormat = "N1")]
-        public float ScienceConst { get; private set; } = 0.5f;
+        public float ConstScience { get; private set; } = 0.5f;
 
         public override void SetDifficultyPreset(GameParameters.Preset preset)
         {
-            
+
             switch (preset)
             {
-                case GameParameters.Preset.Easy:   CareerVessel = 1; ScienceVessel = 0.5f; break;
-                case GameParameters.Preset.Normal: CareerVessel = 2; ScienceVessel = 1; break;
-                case GameParameters.Preset.Hard:   CareerVessel = 5; ScienceVessel = 2;  break;
+                case GameParameters.Preset.Easy:   RelCost = 1; RelScience = 0.5f; break;
+                case GameParameters.Preset.Normal: RelCost = 2; RelScience = 1; break;
+                case GameParameters.Preset.Hard:   RelCost = 5; RelScience = 2;  break;
             }
         }
 
         public override bool Enabled(MemberInfo member, GameParameters parameters)
         {
+            if (member == null) return false;
+
+            if (member.Name == "ConstCost")
+                return ConstCostEnabled;
+
+            if (member.Name == "ConstScience")
+                return ConstScienceEnabled;
+
+            if (member.Name == "RelCost")
+                return RelCostEnabled;
+
+            if (member.Name == "RelScience")
+                return RelScienceEnabled;
+
             return true;
         }
 
         public override bool Interactible(MemberInfo member, GameParameters parameters)
         {
-            if (member == null) return false;
+            if (member == null) 
+                return false;
 
-            if (member.Name == "IgnoreSPH"
-                || member.Name == "ScienceVessel" || member.Name == "CareerVessel"
-                || member.Name == "ScienceIsConst" || member.Name == "CareerIsConst"
-                )
+            if (member.Name == "Enable")
+                return true;
+            else
                 return Enable;
-
-            if (member.Name == "CareerConst")
-                return CareerIsConst && Enable;
-
-            if (member.Name == "ScienceConst")
-                return ScienceIsConst && Enable;
-
-            return true;
         }
 
         public override IList ValidValues(MemberInfo member)
@@ -96,7 +110,7 @@ namespace KVASSNS
     public class KVASSPlanSettings : GameParameters.CustomParameterNode
     {
         public override string Title { get { return Localizer.Format("#KVASS_plan_title"); } }
-        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.CAREER | GameParameters.GameMode.SCIENCE; } }
+        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.NOTMISSION; } }
         public override string Section { get { return "KVASS"; } }
         public override string DisplaySection { get { return "KVASS"; } }
         public override int SectionOrder { get { return 2; } }
@@ -137,10 +151,18 @@ namespace KVASSNS
 
             if (member.Name == "Enable")
                 return true;
+
+            return Enable;
+        }
+
+        public override bool Enabled(MemberInfo member, GameParameters parameters)
+        {
+            if (member == null) return false;
+
             if (member.Name == "QueueAppend" || member.Name == "QueuePrepend")
-                return Queue && Enable;
+                return Queue;
             else
-                return Enable;
+                return true;
         }
 
         public override IList ValidValues(MemberInfo member)
@@ -172,7 +194,7 @@ namespace KVASSNS
     public class KVASSPlanSettings2 : GameParameters.CustomParameterNode
     {
         public override string Title { get { return Localizer.Format("#KVASS_plan_title"); } }
-        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.CAREER | GameParameters.GameMode.SCIENCE; } }
+        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.NOTMISSION; } }
         public override string Section { get { return "KVASS"; } }
         public override string DisplaySection { get { return "KVASS"; } }
         public override int SectionOrder { get { return 3; } }
@@ -180,13 +202,19 @@ namespace KVASSNS
 
         internal static bool interactible = true;
 
-        [GameParameters.CustomIntParameterUI("#KVASS_plan_career_seconds", gameMode = GameParameters.GameMode.CAREER,
-            minValue = 0, maxValue = 180, stepSize = 1)]
-        public int CareerSeconds { get; private set; } = 10;
+        [GameParameters.CustomParameterUI("#KVASS_plan_cost_seconds_enable", toolTip = "#KVASS_plan_cost_seconds_enable_tooltip")]
+        public bool SecondsPerFundEnable { get; private set; } = true;
 
-        [GameParameters.CustomIntParameterUI("#KVASS_plan_science_seconds", gameMode = GameParameters.GameMode.SCIENCE,
+        [GameParameters.CustomIntParameterUI("#KVASS_plan_cost_seconds", 
             minValue = 0, maxValue = 180, stepSize = 1)]
-        public int ScienceSeconds { get; private set; } = 4;
+        public int SecondsPerFund { get; private set; } = 10;
+
+        [GameParameters.CustomParameterUI("#KVASS_plan_mass_seconds_enable", toolTip = "#KVASS_plan_mass_seconds_enable_tooltip")]
+        public bool SecondsPerKgEnable { get; private set; } = false;
+
+        [GameParameters.CustomIntParameterUI("#KVASS_plan_mass_seconds", 
+            minValue = 0, maxValue = 180, stepSize = 1)]
+        public int SecondsPerKg { get; private set; } = 4;
 
         [GameParameters.CustomParameterUI("#KVASS_plan_enable_calendar", toolTip = "#KVASS_plan_enable_calendar_tooltip")]
         public bool CalendarSpeedUp { get; private set; } = true;
@@ -218,10 +246,12 @@ namespace KVASSNS
             minValue = 3, maxValue = 20, stepSize = 1)]
         public int KerbToNextLevel { get; private set; } = 7;
 
-        [GameParameters.CustomParameterUI("#KVASS_plan_enable_sci", toolTip = "#KVASS_plan_enable_sci_tooltip")]
+        [GameParameters.CustomParameterUI("#KVASS_plan_enable_sci", toolTip = "#KVASS_plan_enable_sci_tooltip",
+            gameMode = GameParameters.GameMode.CAREER | GameParameters.GameMode.SCIENCE)]
         public bool SciSpeedUp { get; private set; } = false;
 
-        [GameParameters.CustomIntParameterUI("#KVASS_plan_sci", toolTip = "#KVASS_plan_sci_tooltip", 
+        [GameParameters.CustomIntParameterUI("#KVASS_plan_sci", toolTip = "#KVASS_plan_sci_tooltip",
+            gameMode = GameParameters.GameMode.CAREER | GameParameters.GameMode.SCIENCE,
             minValue = 100, maxValue = 10000, stepSize = 100)]
         public int SciToNextLevel { get; private set; } = 2500;
 
@@ -237,41 +267,47 @@ namespace KVASSNS
             switch (preset)
             {
                 case GameParameters.Preset.Easy:   
-                    CareerSeconds = 10; ScienceSeconds = 4; ConstTime = false; break;
+                    SecondsPerFund = 10; SecondsPerKg = 4; ConstTime = false; break;
                 case GameParameters.Preset.Normal: 
-                    CareerSeconds = 10; ScienceSeconds = 4; break;
+                    SecondsPerFund = 10; SecondsPerKg = 4; break;
                 case GameParameters.Preset.Hard:   
-                    CareerSeconds = 20; ScienceSeconds = 8; break;
+                    SecondsPerFund = 20; SecondsPerKg = 8; break;
             }
         }
 
         public override bool Enabled(MemberInfo member, GameParameters parameters)
-        {   
+        {
+            if (member == null) return false;
+
+            if (member.Name == "SecondsPerFund")
+                return SecondsPerFundEnable;
+
+            if (member.Name == "SecondsPerKg")
+                return SecondsPerKgEnable;
+
+            if (member.Name == "CalendarYearsToNextLevel")
+                return CalendarSpeedUp;
+
+            if (member.Name == "CalendarYearsSpeedUpsMaxCount")
+                return CalendarSpeedUp;
+
+            if (member.Name == "RepToNextLevel")
+                return RepSpeedUp;
+
+            if (member.Name == "KerbToNextLevel")
+                return KerbSpeedUp;
+
+            if (member.Name == "SciToNextLevel")
+                return SciSpeedUp;
+
+            if (member.Name == "ConstTimeDays")
+                return ConstTime;
+
             return true;
         }
 
         public override bool Interactible(MemberInfo member, GameParameters parameters)
         {
-            if (member == null) return false;
-
-            if (member.Name == "CalendarYearsToNextLevel")
-                return CalendarSpeedUp && interactible;
-
-            if (member.Name == "CalendarYearsSpeedUpsMaxCount")
-                return CalendarSpeedUp && interactible;
-
-            if (member.Name == "RepToNextLevel")
-                return RepSpeedUp && interactible;
-
-            if (member.Name == "KerbToNextLevel")
-                return KerbSpeedUp && interactible;
-
-            if (member.Name == "SciToNextLevel")
-                return SciSpeedUp && interactible;
-
-            if (member.Name == "ConstTimeDays")
-                return ConstTime && interactible;
-
             return interactible;
         }
 
@@ -280,6 +316,4 @@ namespace KVASSNS
             return null;
         }
     }
-
-   
 }